requirements.txt
# Auto Consolidator - Requirements File
# Install these packages before building the executable
# Run: pip install -r requirements.txt

# Core dependencies
pandas>=1.3.0
openpyxl>=3.0.0
ttkthemes>=3.2.0
Pillow>=8.0.0

# Build dependencies
pyinstaller>=4.0
